infrastructure/requirements.txt
# AWS CDK Infrastructure Dependencies
aws-cdk-lib>=2.100.0
constructs>=10.0.0
aws-cdk.aws-lambda-python-alpha>=2.100.0a0

# Additional CDK constructs
cdk-nag>=2.27.0
aws-cdk.aws-apigatewayv2-alpha>=2.100.0a0
aws-cdk.aws-apigatewayv2-integrations-alpha>=2.100.0a0

# Python utilities for CDK
boto3>=1.34.0
pyyaml>=6.0.1